if json exists do not create file

+++ statis.sh
@@ -16,10 +16,11 @@ hm=`perl -MPOSIX -le 'print strftime "%H%M",localtime(time-(60))'`
 outJsonFile=$outDir/$ymd$hm.json
 
 #echo $outJsonFile
-
-ihslogfile=443_ssl_access_$ymd.log
-tail -10000 $logdir/$ihslogfile | grep "$hmColon" |awk '{sum[$1]++}END{print "{";for(i in sum) print "\""i"\":"sum[i]",";print "\"yyyyMMddhhmm\":\"'$ymd$hm'\"}"}' >$outJsonFile.tmp
-mv $outJsonFile.tmp $outJsonFile
-chmod 755  "$outJsonFile"
+if [ ! -f "$outJsonFile" ]; then
+    ihslogfile=443_ssl_access_$ymd.log
+    tail -10000 $logdir/$ihslogfile | grep "$hmColon" |awk '{sum[$1]++}END{print "{";for(i in sum) print "\""i"\":"sum[i]",";print "\"yyyyMMddhhmm\":\"'$ymd$hm'\"}"}' >$outJsonFile.tmp
+    mv $outJsonFile.tmp $outJsonFile
+    chmod 755  "$outJsonFile"
+fi
 rm -f $outDir/$yestoday*.json
 #cat $outJsonFile
